requirements.txt
accelerate==0.25.0
agent-executor==0.1.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
altair==5.2.0
annotated-types==0.6.0
anthropic==0.3.11
antlr4-python3-runtime==4.9.3
anyio==3.7.1
arxiv==2.1.0
asgiref==3.7.2
assemblyai==0.20.2
asttokens==2.4.1
attrs==23.1.0
audio-recorder-streamlit==0.0.8
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
bitsandbytes==0.41.3.post2
blinker==1.7.0
boto3==1.34.34
botocore==1.34.34
Brotli==1.1.0
bs4==0.0.1
build==1.0.3
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
clarifai==10.0.0
clarifai-grpc==10.0.3
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
contextlib2==21.6.0
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
dataclasses-json==0.6.3
datasets==2.16.0
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
deta==1.2.0
dill==0.3.7
distro==1.8.0
duckduckgo-search==3.9.11
effdet==0.4.1
emoji==2.10.1
entrypoints==0.4
evaluate==0.4.1
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.4
Faker==23.2.1
fastapi==0.108.0
gizmo-agent==0.1.0
google-api-core==2.15.0
google-auth==2.26.2
google-cloud-aiplatform==1.42.1
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.1
google-cloud-speech==2.23.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
google-search-results==2.4.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
httpx==0.26.0
httpx-oauth==0.13.2
httpx-sse==0.4.0
huggingface-hub==0.20.1
humanfriendly==10.0
humanize==4.9.0
hyperframe==6.0.1
idna==3.6
imageio==2.33.1
importlib-metadata==6.10.0
importlib-resources==6.1.1
inquirerpy==0.3.4
iopath==0.1.10
ipykernel==6.28.0
ipython==8.19.0
kubernetes==29.0.0
langchain==0.1.6
langchain-cli==0.0.21
langchain-community==0.0.19
langchain-core==0.1.23
langchain-experimental==0.0.37
langchain-openai==0.0.2.post1
langchainhub==0.1.14
langdetect==1.0.9
langserve==0.0.41
langsmith==0.0.87
layoutparser==0.3.4
lazy_loader==0.3
llama-index==0.9.30
marshmallow==3.20.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.0
merkle-json==1.0.0
millify==0.1.1
mmh3==4.1.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msal==1.26.0
msal-extensions==1.0.0
multidict==6.0.4
multiprocess==0.70.15
munch==4.0.0
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==1.12.0
opencv-python==4.9.0.80
pandas==2.1.4
parso==0.8.3
pdf2image==1.17.0
pickleshare==0.7.5
Pillow==10.1.0
pinecone-client==3.0.3
streamlit==1.31.1
streamlit-oauth==0.1.8
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchvision==0.16.2+cu121
transformers==4.36.2
trulens-eval==0.23.0
tqdm==4.66.1